refactor(media): move image URL extraction into MediaService and document architecture rules
Remove MediaImageUrlExtractor and consolidate its logic into
MediaService.syncMediaForPost() to comply with the application layer
convention — only *Service classes should be depended on by other
domains. Document Facade/Service/InternalService dependency and return
type rules in backend/README.md and CLAUDE.md.
