+++ backend/blog-api/src/main/kotlin/com/contentria/api/post/application/PostFacade.kt
@@ -2,7 +2,6 @@ package com.contentria.api.post.application
 
 import com.contentria.api.blog.application.BlogService
 import com.contentria.api.category.application.CategoryService
-import com.contentria.api.media.application.MediaImageUrlExtractor
 import com.contentria.api.media.application.MediaService
 import com.contentria.api.post.application.dto.CreateNewPostCommand
 import com.contentria.api.post.application.dto.CreateNewPostInfo
@@ -29,8 +28,7 @@ class PostFacade(
     private val categoryService: CategoryService,
     private val userService: UserService,
     private val markdownService: MarkdownService,
-    private val mediaService: MediaService,
-    private val mediaImageUrlExtractor: MediaImageUrlExtractor
+    private val mediaService: MediaService
 ) {
     @Transactional(readOnly = true)
     fun getPostDetail(blogSlug: String, postSlug: String): PostDetailInfo {
@@ -108,8 +106,7 @@ class PostFacade(
             status = command.status
         )
 
-        val imageUrls = mediaImageUrlExtractor.extractImageUrls(command.contentMarkdown)
-        mediaService.linkMediaToPost(savedPost.id!!, imageUrls)
+        mediaService.syncMediaForPost(savedPost.id!!, command.contentMarkdown)
 
         log.info { "Post created: postId=${savedPost.id}, blogId=${command.blogId}, userId=$userId" }
         return CreateNewPostInfo.from(savedPost)
@@ -133,15 +130,7 @@ class PostFacade(
             status = command.status
         )
 
-        val currentImageUrls = mediaImageUrlExtractor.extractImageUrls(command.contentMarkdown)
-        val previousMedia = mediaService.findByPostId(command.postId)
-        val previousUrls = previousMedia.map { it.publicUrl }
-
-        val newUrls = currentImageUrls.filter { it !in previousUrls }
-        val removedUrls = previousUrls.filter { it !in currentImageUrls }
-
-        mediaService.linkMediaToPost(command.postId, newUrls)
-        mediaService.unlinkMediaFromPost(command.postId, removedUrls)
+        mediaService.syncMediaForPost(command.postId, command.contentMarkdown)
 
         log.info { "Post updated: postId=${updatedPost.id}, blogId=${command.blogId}, userId=$userId" }
         return UpdatePostInfo.from(updatedPost)
